Fix histology/glyph slice reversal bug
- Histology and glyph slices could not be properly reversed when the
reverse_slices option was specified. The reason for this was that the
model took the template string and directly put the slice number into it.
Fix was to create new 'histology_slice' and 'glyph_slice' tables to
properly predefine the slice number in the template string.
- Replaced the 'mr_axis' table with the more general 'axis' table.
- Added 'subject_volume_axis' table to keep the starting and ending
ranges of the slices in each volume. These ranges are specified in the
subject's config.yaml.